Fix a misundestanding with advanced reboot
I've had people ask me why the advanced reboot menu when selecting "hide on lock screen"
was still showing up and even label it as broken. After looking into it, I found that is
not broken but just a misunderstanding. This is only for secured lockscreens so let's change
the string to show this.

Change-Id: I90f0ea477b0aff92d5c9288c21016cf098e000ad