Temporary remove navigation options on devices with hw keys
Having options for the navigation bar such as stock settings and pulse
does not make sense on devices with just hw keys. We can't use these
options until we integrate a method to enable/disable the navigation on
the go.

On devices with a navigation bar, this is will still dependent on our
check for categories.

Change-Id: Iba2d96889544f9e83b3106aeff6151327d55a39c